despliegue_3/requirements.txt
streamlit==1.55.0
pypdf==6.9.2
pillow==12.1.1
openpyxl==3.1.5